fix(core): #2337 review-4 — move appendTextFallbackForNonObject to leaf wire/textFallback.ts (break value cycle)
