+++ packages/core/src/wire/codec.ts
@@ -328,20 +328,3 @@ export function isSpecNotificationMethod(method: string): boolean {
 }
 
 const ALL_CODECS: readonly WireCodec[] = [rev2025Codec, rev2026Codec];
-
-/**
- * SEP-2106 §4.3 TextContent auto-append, era-agnostic, called from BOTH
- * codecs' {@link WireCodec.projectCallToolResult}: when `structuredContent`
- * is a non-object value (array/primitive/`null`) and the handler authored no
- * `type:'text'` block, append `{type:'text', text: JSON.stringify(value)}`.
- * Object-shaped (or absent) `structuredContent` returns the same reference.
- */
-export function appendTextFallbackForNonObject(result: CallToolResult): CallToolResult {
-    const sc = result.structuredContent;
-    if (sc === undefined) return result;
-    const isNonObjectValue = typeof sc !== 'object' || sc === null || Array.isArray(sc);
-    if (!isNonObjectValue) return result;
-    const hasTextContent = result.content?.some(c => c.type === 'text') ?? false;
-    if (hasTextContent) return result;
-    return { ...result, content: [...(result.content ?? []), { type: 'text' as const, text: JSON.stringify(sc) }] };
-}
