+++ packages/core/src/wire/rev2025-11-25/codec.ts
@@ -30,7 +30,7 @@ import type * as z from 'zod/v4';
 
 import type { CallToolResult, Result } from '../../types/types.js';
 import type { DecodedResult, EnvelopeIssue, LiftedWireMaterial, OutboundEnvelopeMaterial, ValidateOutcome, WireCodec } from '../codec.js';
-import { appendTextFallbackForNonObject } from '../codec.js';
+import { appendTextFallbackForNonObject } from '../textFallback.js';
 import { isNonObjectJsonSchemaRoot, wrapOutputSchemaForLegacy } from './legacyWrap.js';
 import { getNotificationSchema, getRequestSchema, getResultSchema, hasNotificationMethod2025, hasRequestMethod2025 } from './registry.js';
 import { CreateMessageResultSchema, CreateMessageResultWithToolsSchema } from './schemas.js';
